+++ enlighten/measurement/MeasurementsFeature.py
@@ -1,3 +1,4 @@
+import threading
 import logging
 import json
 import csv
@@ -14,6 +15,7 @@ from enlighten import common
 from enlighten.common import msgbox
 from enlighten.EnlightenFeature import EnlightenFeature
 from enlighten.measurement.Measurement import Measurement
+from enlighten.file_io.EnlightenJSONEncoder import EnlightenJSONEncoder
 
 if common.use_pyside2():
     from PySide2 import QtWidgets
@@ -21,6 +23,17 @@ else:
     from PySide6 import QtWidgets
 
 log = logging.getLogger(__name__)
+
+class ExportWorker(threading.Thread):
+
+    def __init__(self, measurements):
+        threading.Thread.__init__(self)
+        self.measurements = measurements
+
+    def run(self):
+        log.debug("ExportWorker: calling perform_export_from_worker")
+        self.measurements.perform_export_from_worker()
+        log.debug("ExportWorker: back from perform_export_from_worker")
 
 ##
 # This class represents the set of Measurement objects which have been saved
@@ -68,6 +81,8 @@ class MeasurementsFeature(EnlightenFeature):
         cfu.verticalLayout_scope_capture_save.addItem(spacer)
 
         self.update_count()
+
+        self.export_worker = None
 
     # ##########################################################################
     #                                                                          #
@@ -317,20 +332,19 @@ class MeasurementsFeature(EnlightenFeature):
     # @param filename: BatchCollection generates one so we needn't prompt user
     # @param prompt: prompt for verification (False for unattended operation)
     def export_session(self, filename=None, prompt=True):
+        self.export_filename = filename
+        self.export_directory = common.get_default_data_dir()
 
         if not self.count():
             log.warn("no measurements to export")
             return
 
-        # by default, all Sessions are stored in ~/EnlightenSpectra
-        directory = common.get_default_data_dir()
-
         ########################################################################
         # Generate pathname
         ########################################################################
 
-        visible_only = False
-        if filename is None:
+        self.export_visible_only = False
+        if self.export_filename is None:
             now = datetime.now()
 
             default_filename = f"{self.ctl.save_options.prefix()}-" if self.ctl.save_options.has_prefix() else "Session-"
@@ -340,7 +354,7 @@ class MeasurementsFeature(EnlightenFeature):
             default_filename = self.measurements[-1].expand_template(default_filename)
 
             if not prompt:
-                filename = default_filename
+                self.export_filename = default_filename
             else:
                 # prompt the user to override the default filename
                 result = self.ctl.gui.msgbox_with_lineedit_and_checkbox(
@@ -350,54 +364,67 @@ class MeasurementsFeature(EnlightenFeature):
                     extra_button_label = "Browse",
                     checkbox_text = "Only export displayed traces")
                 log.debug(f"msgbox result: {result}")
-                filename = result["lineedit"]
+                self.export_filename = result["lineedit"]
                 visible_only = result["checked"]
                 if not result["ok"]:
                     log.info("cancelling export")
                     return
 
                 # if the user clicked 'Browse', let them navigate to where they want the export saved [#420]
-                if result["extra_button_clicked"] or not filename:
-                    pathname = self.ctl.file_manager.save_dialog(filename=filename, caption="Select where to save your exported measurements")
-                    directory, filename = os.path.split(pathname)
+                if result["extra_button_clicked"] or not self.export_filename:
+                    pathname = self.ctl.file_manager.save_dialog(filename=self.export_filename, caption="Select where to save your exported measurements")
+                    self.export_directory, self.export_filename = os.path.split(pathname)
 
-                if not filename:
+                if not self.export_filename:
                     log.info("cancelling export")
                     return
 
         # warn user if they are about to overwrite an existing file
         # it looks like only risk is for .csv and .json extensions
-        file_exists = (os.path.exists(os.path.join(directory, f'{filename}.csv'))
-                       or os.path.exists(os.path.join(directory, f'{filename}.json'))
-                       or os.path.exists(os.path.join(directory, f'{filename}.spc')))
+        file_exists = (os.path.exists(os.path.join(self.export_directory, f'{self.export_filename}.csv'))
+                       or os.path.exists(os.path.join(self.export_directory, f'{self.export_filename}.json'))
+                       or os.path.exists(os.path.join(self.export_directory, f'{self.export_filename}.spc')))
 
-        log.info(f"checking for: {os.path.join(directory, filename)}, file exists={file_exists}")
+        log.info(f"checking for: {os.path.join(self.export_directory, self.export_filename)}, file exists={file_exists}")
 
         if file_exists:
-            should_overwrite = msgbox(prompt=f"Do you wish to overwrite the existing file: {filename}?",
+            should_overwrite = msgbox(prompt=f"Do you wish to overwrite the existing file: {self.export_filename}?",
                                       informative_text="All data in the previous file will be lost.",
                                       buttons="Yes|No") == "Yes"
         else:
             should_overwrite = False
 
         if not file_exists or should_overwrite:
-            # doesn't use the dict
-            if self.ctl.save_options.save_csv():
-                self.export_session_csv(directory, filename, visible_only=visible_only)
+            self.export_worker = ExportWorker(self)
+            self.export_worker.setDaemon(True)
+            self.export_worker.start()
 
-            # cache export dictionary so we can re-use it between JSON and ExternalAPI
-            if self.ctl.save_options.save_json() or "export" in self.observers:
-                export = self.generate_export_dict(visible_only=visible_only)
+    def perform_export_from_worker(self):
+        """ Note this is called from ExportWorker's Python thread, not the Qt GUI thread...should be okay? """
 
-            if self.ctl.save_options.save_json():
-                self.export_session_json(directory, filename, export)
+        log.debug("perform_export_from_worker: start")
 
-            if self.ctl.save_options.save_spc():
-                self.export_session_spc(directory, filename, visible_only=visible_only)
+        if self.ctl.save_options.save_csv():
+            self.export_session_csv(self.export_directory, self.export_filename, visible_only=self.export_visible_only)
+
+        # cache export dictionary so we can re-use it between JSON and ExternalAPI
+        if self.ctl.save_options.save_json() or "export" in self.observers:
+            export = self.generate_export_dict(visible_only=self.export_visible_only)
+
+        if self.ctl.save_options.save_json():
+            self.export_session_json(self.export_directory, self.export_filename, export)
+
+        if self.ctl.save_options.save_spc():
+            self.export_session_spc(self.export_directory, self.export_filename, visible_only=self.export_visible_only)
+
+        self.ctl.progress_bar.hide()
 
         # MZ: this seems weird...why are we sending visible_only as the value, 
-        # instead of 'export' (the dict)? Who subscribes to this?
-        self.notify_observers_with_value(visible_only, "export")
+        # instead of 'export' (the dict)? Who subscribes to this? A: NOBODY?
+        #
+        # self.notify_observers_with_value(self.export_visible_only, "export")
+
+        log.debug("perform_export_from_worker: done")
 
     def read_measurements(self):
         return self.generate_export_dict()
@@ -478,15 +505,20 @@ class MeasurementsFeature(EnlightenFeature):
             filename += ".json"
         pathname = os.path.join(directory, filename)
 
+        self.ctl.marquee.info("exporting %d spectra to JSON..." % len(export))
+        self.ctl.progress_bar.show_continuous()
+
         # log.debug("traversing 'export' to look for non-exportable blocks...")
         # util.traverse_json(export)
 
-        s = json.dumps(export, sort_keys=True, indent=2, default=lambda o: o.to_json())
-        s = util.clean_json(s, reformat=False)
+        s = json.dumps(export, cls=EnlightenJSONEncoder, sort_keys=True, indent=2, default=lambda o: o.to_json())
+        s = util.clean_json(s)
 
         log.debug(f"export_session_json: writing {pathname}")
         with open(pathname, "w") as f:
             f.write(s)
+
+        self.ctl.marquee.info("exported %d spectra to JSON" % len(export))
 
     def export_session_csv(self, directory, filename, visible_only=False):
         if not filename.endswith(".csv"):
@@ -501,6 +533,7 @@ class MeasurementsFeature(EnlightenFeature):
             export_measurements = self.measurements
 
         log.info("exporting %d measurements in %s order to %s", len(export_measurements), order, pathname)
+        self.ctl.marquee.info("exporting %d spectra to CSV" % len(export_measurements))
 
         ########################################################################
         # Generate the export
@@ -514,7 +547,7 @@ class MeasurementsFeature(EnlightenFeature):
                 else:
                     self.export_by_column(csv_writer, visible_only)
 
-            self.ctl.marquee.info("exported %d spectra" % len(export_measurements))
+            self.ctl.marquee.info("exported %d spectra to CSV" % len(export_measurements))
             log.info("exported %d measurements in %s order to %s", self.count(), order, pathname)
 
         except Exception:
@@ -931,6 +964,8 @@ class MeasurementsFeature(EnlightenFeature):
                             row.append(get_pr_header_value(m, header, pixel, pr=interpolated[m]))
                 csv_writer.writerow(row)
 
+                self.ctl.progress_bar.set(100 * pixel / self.ctl.interp.total_pixels())
+
         else:
 
             #####################################################################
@@ -973,6 +1008,7 @@ class MeasurementsFeature(EnlightenFeature):
                             row.extend(BLANK * len(pr_headers))
 
                 csv_writer.writerow(row)
+                self.ctl.progress_bar.set(100 * pixel / max_pixels)
 
     ##
     # In the row-based export, try to follow historical Dash conventions, as