+++ enlighten/ui/ProgressBarFeature.py
@@ -18,31 +18,64 @@ class ProgressBarFeature(EnlightenFeature):
         cfu = ctl.form.ui
         self.pb = cfu.readingProgressBar
 
-        self.hide_timer = QtCore.QTimer()
-        self.hide_timer.setSingleShot(True)
-        self.hide_timer.timeout.connect(self.hide)
+        self.visible = False
+        self.value = 0
+
+        self.timer = QtCore.QTimer()
+        self.timer.setSingleShot(True)
+        self.timer.timeout.connect(self.tick)
+        self.timer.start(1000)
+
+    def tick(self):
+        # default to 1Hz
+        next_check_ms = 1000
+
+        self.visible = self.value < 0 or self.value > 0
+
+        log.debug(f"tick: visible {self.visible}, value {self.value}")
+
+        if not self.visible:
+            self.pb.setVisible(False)
+        else:
+            if self.value < 0:
+                # unbounded "busy" animation
+                self.pb.setVisible(True)
+                self.pb.setRange(0, 0)
+            else:
+                # round to int 0-100
+                self.pb.setRange(0, 100)
+                self.value = int(min(100, max(0, round(self.value, 0))))
+                
+                if self.value == 0:
+                    # hide immediately on zero
+                    self.visible = False
+                    self.pb.setVisible(False)
+                else:
+                    self.pb.setValue(self.value)
+                    self.pb.setVisible(True)
+
+                    if self.value >= 100:
+                        # hide on the NEXT tick
+                        log.debug("tick: hide on next")
+                        self.value = 0
+                    else: 
+                        # tick at 4Hz above 1%, below 100%
+                        next_check_ms = 250
+
+        self.timer.start(next_check_ms)
+
+    def show_continuous(self):
+        log.debug("show continuous")
+        self.set(-1)
 
     def hide(self):
-        self.pb.setVisible(False)
+        log.debug("hiding")
+        self.set(0)
 
     def set(self, value):
-        """ Hide at zero. Set timer to disappear 3sec after reaching 100. """
-
-        if value < 0:
-            # unbounded "busy" animation
-            self.pb.setVisible(True)
-            self.pb.setRange(0, 0)
-        else:
-            # round to int 0-100
-            self.pb.setRange(0, 100)
-            value = int(min(100, max(0, round(value, 0))))
-            log.debug(f"setting {value}")
-            
-            if value == 0:
-                self.pb.setVisible(False)
-            else:
-                self.pb.setValue(value)
-                self.pb.setVisible(True)
-
-                if value >= 100:
-                    self.hide_timer.start(1000)
+        """ 
+        Can be called from outside GUI thread.
+        New settings get applied on next tick.
+        """
+        log.debug(f"set: {value}")
+        self.value = value